Moved from namespace FacturaScripts\Core\Base\Utils to FacturaScripts\Core\Base
